move AssertionGeneratorUtils higher up in folder hierarchy
