+++ dspot/src/main/java/eu/stamp_project/dspot/assertgenerator/components/AssertionRemover.java
@@ -1,6 +1,6 @@
 package eu.stamp_project.dspot.assertgenerator.components;
 
-import eu.stamp_project.dspot.assertgenerator.components.utils.AssertionGeneratorUtils;
+import eu.stamp_project.dspot.assertgenerator.utils.AssertionGeneratorUtils;
 import eu.stamp_project.test_framework.TestFramework;
 import eu.stamp_project.utils.CloneHelper;
 import spoon.reflect.code.*;
