+++ dspot/src/main/java/eu/stamp_project/dspot/assertgenerator/components/testmethodreconstructor/assertionsyntaxbuilder/InvocationSetup.java
@@ -1,6 +1,6 @@
 package eu.stamp_project.dspot.assertgenerator.components.testmethodreconstructor.assertionsyntaxbuilder;
 
-import eu.stamp_project.dspot.assertgenerator.components.utils.AssertionGeneratorUtils;
+import eu.stamp_project.dspot.assertgenerator.utils.AssertionGeneratorUtils;
 import eu.stamp_project.test_framework.TestFramework;
 import eu.stamp_project.test_framework.assertions.AssertEnum;
 import spoon.reflect.code.CtExpression;
