+++ dspot/src/main/java/eu/stamp_project/test_framework/implementations/junit/JUnitSupport.java
@@ -1,7 +1,7 @@
 package eu.stamp_project.test_framework.implementations.junit;
 
 import eu.stamp_project.compare.ObjectLog;
-import eu.stamp_project.dspot.assertgenerator.components.utils.AssertionGeneratorUtils;
+import eu.stamp_project.dspot.assertgenerator.utils.AssertionGeneratorUtils;
 import eu.stamp_project.test_framework.AbstractTestFramework;
 import eu.stamp_project.test_framework.assertions.AssertEnum;
 import eu.stamp_project.testrunner.runner.Failure;
